真正的温度传入&修改&写入&汇报

+++ serves/sensors.go
@@ -18,7 +18,7 @@ func BodyRes(resflag int) {
 
 // TempRes -serves 温度传感器
 // 传入温度值 float
-func TempRes(temp float64) {
+func TempRes(temp float32) {
 	if temp < 100.0 && temp > -20.0 {
 		sqls.TempRes(temp)
 	} else {
@@ -28,7 +28,7 @@ func TempRes(temp float64) {
 
 // HumiRes -serves 湿度传感器
 // 传入湿度百分比
-func HumiRes(humi float64) {
+func HumiRes(humi float32) {
 	if humi < 100 && humi > 0 {
 		sqls.HumiRes(humi)
 	} else {
